+++ src/plugins/footnotes/footnotes-utils.ts
@@ -14,20 +14,23 @@
  * limitations under the License.
  */
 
-import { InlineFootnoteNode, schema } from '@manuscripts/transform'
-import { ResolvedPos } from 'prosemirror-model'
-import { EditorState, Transaction } from 'prosemirror-state'
-import { ReplaceStep } from 'prosemirror-transform'
+import { FootnoteWithIndex } from '@manuscripts/style-guide'
+import { skipTracking } from '@manuscripts/track-changes-plugin'
 import {
+  InlineFootnoteNode,
+  isFootnoteNode,
+  ManuscriptNode,
+  schema,
+} from '@manuscripts/transform'
+import { Fragment, ResolvedPos } from 'prosemirror-model'
+import { Transaction } from 'prosemirror-state'
+import {
+  ContentNodeWithPos,
   findChildren,
   findChildrenByType,
   findParentNodeClosestToPos,
+  NodeWithPos,
 } from 'prosemirror-utils'
-
-import {
-  buildTableFootnoteLabels,
-  orderTableFootnotes,
-} from '../../lib/footnotes-utils'
 
 export const findTableInlineFootnoteIds = ($pos: ResolvedPos) => {
   const tableElement = findParentNodeClosestToPos(
@@ -47,44 +50,93 @@ export const findTableInlineFootnoteIds = ($pos: ResolvedPos) => {
   )
 }
 
-export const updateTableInlineFootnoteLabels = (
-  transactions: readonly Transaction[],
-  oldState: EditorState,
-  newState: EditorState
+export const getNewFootnotePos = (
+  footnotesElement: NodeWithPos,
+  footnoteIndex: number
 ) => {
-  const tableInlineFootnoteChange = transactions.find((tr) =>
-    tr.steps.find((s) => {
-      if (s instanceof ReplaceStep) {
-        const $pos = oldState.doc.resolve((s as ReplaceStep).from)
-        return (
-          $pos.node().type === schema.nodes.table_cell &&
-          $pos.node($pos.depth - 2).type === schema.nodes.table
-        )
+  let newFootnotePos = footnotesElement.pos + footnotesElement.node.nodeSize - 1
+
+  footnotesElement.node.descendants((node, pos, parent, index) => {
+    if (isFootnoteNode(node)) {
+      if (footnoteIndex === ++index) {
+        newFootnotePos = footnotesElement.pos + pos + (index === 1 ? 2 : 1)
+      }
+    }
+  })
+
+  return newFootnotePos
+}
+
+export const buildTableFootnoteLabels = (node: ManuscriptNode) => {
+  const labels = new Map<string, string | undefined>(
+    findChildrenByType(node, schema.nodes.footnote).map((node) => [
+      node.node.attrs.id,
+      undefined,
+    ])
+  )
+
+  let index = 0
+
+  findChildrenByType(node, schema.nodes.inline_footnote)
+    .map(({ node }) => node.attrs.rids)
+    .flat()
+    .map((rid: string) => {
+      if (!labels.get(rid)) {
+        labels.set(rid, String(++index))
       }
     })
+
+  return labels
+}
+
+/**
+ *  This will make sure un-cited table footnotes are at the end of the list
+ *  and order cited footnote based on the position of inline footnote
+ */
+export const orderTableFootnotes = (
+  tr: Transaction,
+  footnotesElementWithPos: NodeWithPos,
+  position: number
+) => {
+  const tablesFootnoteLabels = buildTableFootnoteLabels(
+    tr.doc.nodeAt(tr.mapping.map(position)) as ManuscriptNode
   )
 
-  if (!tableInlineFootnoteChange) {
-    return null
-  }
+  const footnotes = findChildrenByType(
+    footnotesElementWithPos.node,
+    schema.nodes.footnote
+  ).map((nodeWithPos) => ({
+    node: nodeWithPos.node,
+    index: tablesFootnoteLabels.get(nodeWithPos.node.attrs.id),
+  })) as FootnoteWithIndex[]
 
-  const step = tableInlineFootnoteChange.steps[0] as ReplaceStep
+  const orderedFootnotes = [...footnotes]
+    .sort((a, b) => {
+      if (a.index !== undefined && b.index !== undefined) {
+        return Number.parseInt(a.index) - Number.parseInt(b.index)
+      } else {
+        return a.index === undefined ? 1 : -1
+      }
+    })
+    .map(({ node }) => node)
 
-  const tr = newState.tr
-  const $pos = newState.doc.resolve(step.from)
-  const table = findParentNodeClosestToPos(
-    $pos,
-    (node) => node.type === schema.nodes.table_element
+  const { node: footnotesElement, pos } = footnotesElementWithPos
+  const footnoteElementPos = position + pos
+
+  return skipTracking(
+    tr.replaceWith(
+      tr.mapping.map(footnoteElementPos),
+      tr.mapping.map(footnoteElementPos + footnotesElement.nodeSize),
+      Fragment.fromArray(orderedFootnotes)
+    )
   )
+}
 
-  const footnotesElementWithPos =
-    table &&
-    findChildrenByType(table.node, schema.nodes.footnotes_element).pop()
-
-  if (!table || !footnotesElementWithPos) {
-    return null
-  }
-
+export const updateTableInlineFootnoteLabels = (
+  tr: Transaction,
+  table: ContentNodeWithPos,
+  footnotesElementWithPos: NodeWithPos
+) => {
   const labels = buildTableFootnoteLabels(table.node)
 
   findChildrenByType(table.node, schema.nodes.inline_footnote).map(
